added close button, dinamic maxWidth, title changed to ReactNode, components seperated to files
