+++ lib/main.ts
@@ -1,4 +1,4 @@
 import "rc-drawer/assets/index.css";
 import Drawer from "./components/Drawer";
-
+export type { ResizableDrawerProps } from "./components/Drawer/types";
 export default Drawer;
